+++ src/Farmer/Common.fs
@@ -2,6 +2,16 @@
 
 open System
 
+type NonEmptyList<'T> =
+    private | NonEmptyList of List<'T>
+    /// Unwraps the inner List contents.
+    member this.Value = match this with NonEmptyList list -> list
+module NonEmptyList =
+    let create list =
+        match list with
+        | [] -> failwith "This list must always have at least one item in it."
+        | list -> NonEmptyList list
+
 [<AutoOpen>]
 module internal DuHelpers =
     let makeAll<'TUnion> =
@@ -53,6 +63,15 @@ module LocationExtensions =
         static member NorwayWest = Location "NorwayWest"
         static member NorwayEast = Location "NorwayEast"
         static member Global = Location "global"
+
+[<AutoOpen>]
+module DataLocationExtensions =
+    type DataLocation with
+        static member AsiaPacific = DataLocation "Asia Pacific"
+        static member Australia = DataLocation "Australia"
+        static member Europe = DataLocation "Europe"
+        static member UnitedKingdom = DataLocation "United Kingdom"
+        static member UnitedStates = DataLocation "United States"
 
 type OS = Windows | Linux
 
@@ -399,11 +418,40 @@ module Storage =
         member this.ResourceName = match this with StorageResourceName name -> name
 
     type DefaultAccessTier = Hot | Cool
-    type StoragePerformance = Standard | Premium
-    type BasicReplication = LRS | ZRS
-    type BlobReplication = LRS | GRS | RAGRS
-    type V1Replication = LRS of StoragePerformance | GRS | RAGRS
-    type V2Replication = LRS of StoragePerformance | GRS | ZRS | GZRS | RAGRS | RAGZRS
+    type StoragePerformance =
+        | Standard | Premium
+        member this.ArmValue = match this with Standard -> "Standard" | Premium -> "Premium"
+    type BasicReplication =
+        | LRS | ZRS
+        member this.ReplicationModelDescription =
+            match this with
+            | LRS -> "LRS"
+            | ZRS -> "ZRS"
+    type BlobReplication =
+        | LRS | GRS | RAGRS
+        member this.ReplicationModelDescription =
+            match this with
+            | LRS -> "LRS"
+            | GRS -> "GRS"
+            | RAGRS -> "RAGRS"
+    type V1Replication =
+        | LRS of StoragePerformance | GRS | RAGRS
+        member this.ReplicationModelDescription =
+            match this with
+            | LRS _ -> "LRS"
+            | GRS -> "GRS"
+            | RAGRS -> "RAGRS"
+    type V2Replication =
+        | LRS of StoragePerformance | GRS | ZRS | GZRS | RAGRS | RAGZRS
+        member this.ReplicationModelDescription =
+            match this with
+            | LRS _ -> "LRS"
+            | GRS -> "GRS"
+            | ZRS -> "ZRS"
+            | GZRS -> "GZRS"
+            | RAGRS -> "RAGRS"
+            | RAGZRS -> "RAGZRS"
+
     type GeneralPurpose = V1 of V1Replication | V2 of V2Replication * DefaultAccessTier option
     type Sku =
         | GeneralPurpose of GeneralPurpose
@@ -439,6 +487,40 @@ module Storage =
 
     /// Represents no filters for a lifecycle rule
     let NoRuleFilters : string list = []
+
+    type AllOrSpecific<'T> =
+        | All
+        | Specific of 'T list
+
+    type HttpMethod =
+        | DELETE | GET | HEAD | MERGE | POST | OPTIONS | PUT | PATCH
+        static member All = NonEmptyList.create [ DELETE; GET; HEAD; MERGE; POST; OPTIONS; PUT; PATCH ]
+        member this.ArmValue =
+            match this with
+            | DELETE -> "DELETE" | GET -> "GET" | HEAD -> "HEAD" | MERGE -> "MERGE"
+            | POST -> "POST" | OPTIONS -> "OPTIONS" | PUT -> "PUT" | PATCH -> "PATCH"
+    type CorsRule =
+        { AllowedOrigins : AllOrSpecific<Uri>
+          AllowedMethods : HttpMethod NonEmptyList
+          MaxAgeInSeconds : int
+          ExposedHeaders : AllOrSpecific<string>
+          AllowedHeaders : AllOrSpecific<string> }
+        static member AllowAll =
+            { AllowedOrigins = All
+              AllowedMethods = HttpMethod.All
+              MaxAgeInSeconds = 0
+              ExposedHeaders = All
+              AllowedHeaders = All }
+        /// Creates a new CORS rule with
+        static member create (?allowedOrigins, ?allowedMethods, ?maxAgeInSeconds, ?exposedHeaders, ?allowedHeaders) =
+            let mapDefault mapper defaultValue = Option.map mapper >> Option.defaultValue defaultValue
+            { AllowedOrigins = allowedOrigins |> mapDefault (List.map Uri >> Specific) CorsRule.AllowAll.AllowedOrigins
+              AllowedMethods = allowedMethods |> mapDefault NonEmptyList.create CorsRule.AllowAll.AllowedMethods
+              MaxAgeInSeconds = defaultArg maxAgeInSeconds CorsRule.AllowAll.MaxAgeInSeconds
+              ExposedHeaders = exposedHeaders |> mapDefault Specific CorsRule.AllowAll.ExposedHeaders
+              AllowedHeaders = allowedHeaders |> mapDefault Specific CorsRule.AllowAll.AllowedHeaders }
+    [<RequireQualifiedAccess>]
+    type StorageService = Blobs | Tables | Files | Queues
 
 module WebApp =
     type WorkerSize = Small | Medium | Large | Serverless