+++ RELEASE_NOTES.md
@@ -2,8 +2,10 @@ Release Notes
 =============
 
 ## 1.5.1
+* Communication Services: add builder.
 * ServiceBus: Fix an issue with Premium Sku ARM Writer
 * ServiceBus: Fix an issue with Rules depends on ARM Writer
+* Storage Accounts: Support for CORS.
 
 ##  1.5.0
 * Container Groups: Support for init containers.